+++ Aula11/zoologico/src/main/java/br/com/wcc/model/Galinha.java
